requirements.txt
streamlit
google-generativeai
pyswisseph